SD 2/ev_charging/api_central/app/requirements.txt
kafka-python==2.0.2
flask